% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_age.R
\name{calc_age}
\alias{calc_age}
\title{Calculate Age}
\usage{
calc_age(
  birth_date,
  to_date = Sys.Date(),
  freq = years,
  set_negative_na = FALSE,
  warn = FALSE
)
}
\arguments{
\item{birth_date}{vector of dates}

\item{to_date}{vector of dates, or date of length 1, default = Sys.Date()}

\item{freq}{one of year, months or weeks. These are functions to pass from
the lubridate package.}

\item{set_negative_na}{if TRUE will change negative ages to NA}

\item{warn}{if TRUE will warn where dates are implausible}
}
\value{
age in years
}
\description{
Calculate age in years based on a birth date and to date
}
