% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_col_names.R
\name{check_col_names}
\alias{check_col_names}
\title{Check Presence of Specified Column Names in a Data Frame}
\usage{
check_col_names(df, col_names, df_name = "df")
}
\arguments{
\item{df}{A data frame to check for specified column names.}

\item{col_names}{A character vector specifying the column names expected to be present in the data frame.}

\item{df_name}{Optional character string specifying the name of the data frame
being checked. This name is used in error messages for clarity.}
}
\value{
\code{TRUE} if the data frame contains all of the specified column names, otherwise
an error is thrown detailing the missing columns.
}
\description{
Validates if a data frame contains all of the specified column names.
An optional \code{df_name} argument enhances error messages with the data frame's name.
}
\examples{
df_example <- data.frame(x = 1:10, y = letters[1:10])
check_col_names(df_example, c("x", "y"), df_name = "df_example") # Columns are present
check_col_names(df_example, c("x", "z"), df_name = "df_example") # Column 'z' is missing

}
