% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_df.R
\name{check_df}
\alias{check_df}
\title{Comprehensive Data Frame Checks}
\usage{
check_df(
  df,
  num_rows = NULL,
  num_cols = NULL,
  col_names = NULL,
  col_types = NULL,
  df_name = "df"
)
}
\arguments{
\item{df}{A data frame to be validated.}

\item{num_rows}{A single positive whole number or a vector of length 2 indicating
the exact, minimum, or maximum number of rows allowed. Use \code{NA} to omit the minimum
or maximum constraint. If \code{NULL}, this check is skipped.}

\item{num_cols}{A single positive whole number or a vector of length 2 indicating
the exact, minimum, or maximum number of columns allowed. Use \code{NA} to omit the minimum
or maximum constraint. If \code{NULL}, this check is skipped.}

\item{col_names}{A character vector specifying the column names expected to be present
in the data frame in the given order. If \code{NULL}, this check is skipped.}

\item{col_types}{A named vector of types/classes expected for the data frame's columns.
If \code{NULL}, this check is skipped.}

\item{df_name}{Optional character string specifying the name of the data frame being checked.
Used in error messages for clarity.}
}
\value{
\code{TRUE} if the data frame passes all specified checks, otherwise
an error is thrown detailing the first encountered discrepancy.
}
\description{
Performs a series of checks on a data frame, including validations for the number of rows,
the number of columns, presence and order of specified column names, and column types.
}
\examples{
df_example <- data.frame(name = as.character(1:10), age = as.integer(21:30))
check_df(df_example, num_rows = c(1, NA), num_cols = 2,
         col_names = c("name", "age"), col_types = c(name = "character", age = "integer"),
         df_name = "df_example")

}
