% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_num_cols.R
\name{check_num_cols}
\alias{check_num_cols}
\title{Check the Number of Columns in a Data Frame}
\usage{
check_num_cols(df, num_cols, df_name = "df")
}
\arguments{
\item{df}{A data frame whose columns are to be checked.}

\item{num_cols}{A single positive whole number or a vector of length 2
indicating the exact, minimum, or maximum number of columns allowed. Use \code{NA}
to omit the minimum or maximum constraint.}

\item{df_name}{Optional character string specifying the name of the data frame
being checked. This name is used in error messages for clarity.}
}
\value{
\code{TRUE} if the data frame meets the column count criteria, otherwise
an error is thrown detailing the discrepancy.
}
\description{
Validates the number of columns in a given data frame against specified criteria.
}
\examples{
df_example <- data.frame(x = 1:10, y = letters[1:10])
check_num_cols(df_example, 2, df_name = "df_example") # Exact number of columns
check_num_cols(df_example, c(1, 3), df_name = "df_example") # Range of columns
check_num_cols(df_example, c(NA, 4), df_name = "df_example") # Maximum number of columns
check_num_cols(df_example, c(2, NA), df_name = "df_example") # Minimum number of columns

}
