% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_csv_to_pqt.R
\name{convert_csv_to_pqt}
\alias{convert_csv_to_pqt}
\title{Convert CSV File to Parquet Format}
\usage{
convert_csv_to_pqt(
  file_in,
  file_out = NA,
  .clean_names = FALSE,
  overwrite = TRUE
)
}
\arguments{
\item{file_in}{A character string specifying the path to the input CSV file.}

\item{file_out}{A character string specifying the path to the output Parquet file.
Defaults to \code{NA}, which will use the same path and name as \code{file_in},
replacing the \code{.csv} extension with \code{.pqt}.}

\item{.clean_names}{logical, determines if input data frame should be passed
to janitor::clean_names}

\item{overwrite}{if TRUE will overwrite any existing file_out}
}
\value{
This function returns \code{invisible(NA)}. The output file is saved as a Parquet file.
}
\description{
This function reads in a CSV file and converts it to Parquet format.
}
\examples{
\dontrun{
# Convert a CSV file to Parquet format with default output file name
convert_csv_to_pqt("data/input_data.csv")

# Convert a CSV file to Parquet format with a specified output file name
convert_csv_to_pqt("data/input_data.csv", "data/output_data.pqt")
}

}
