% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record-linkage.R
\name{generate_pairs}
\alias{generate_pairs}
\title{Generate Record Linkage Pairs}
\usage{
generate_pairs(df1, df2, exclude = numeric(0), blockfld = list())
}
\arguments{
\item{df1}{A data frame containing the first dataset to be compared.}

\item{df2}{A data frame containing the second dataset to be compared.}

\item{exclude}{see RLBigDataLinkage}

\item{blockfld}{see RLBigDataLinkage}
}
\value{
A \code{RecLinkData} object with the computed pairwise comparison weights.
}
\description{
The \code{generate_pairs} function generates possible record linkage pairs between
two data frames using the RecordLinkage package. It computes the pairwise
similarity of records between two datasets and calculates weights for each
pair based on string similarity using the Levenshtein distance.
}
\examples{
# Example usage
df1 <- data.frame(name = c("John", "Jane"), age = c(25, 30))
df2 <- data.frame(name = c("Jon", "Janet"), age = c(26, 31))
pairs <- generate_pairs(df1, df2, excluded_cols = c("age"))

}
