% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{get_time_stamp}
\alias{get_time_stamp}
\title{Generate a Timestamp with Milliseconds}
\usage{
get_time_stamp()
}
\value{
A string representing the current timestamp in the format
"YYYYMMDD-HHMMSSmmm", where \code{mmm} is the milliseconds.
}
\description{
The \code{get_time_stamp} function generates a timestamp string in the format
"YYYYMMDD-HHMMSSmmm", where \code{mmm} represents milliseconds. It captures the
current system time and formats it as a string suitable for file names or
logging purposes, ensuring precision down to milliseconds.
}
\examples{
# Generate a timestamp for the current time
get_time_stamp()

}
