% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{round_nearest}
\alias{round_nearest}
\title{Round to Nearest Multiple}
\usage{
round_nearest(x, k)
}
\arguments{
\item{x}{A numeric vector of values to round.}

\item{k}{The multiple to which the values should be rounded. Must be
a positive number.}
}
\value{
A numeric vector with the values rounded to the nearest multiple of \code{k}.
}
\description{
Rounds a numeric vector to the nearest specified multiple, similar to the
MROUND function in Excel. This function uses mathematical rounding - see
\code{\link{round_number}} for more information on this.
}
\examples{
round_nearest(123, 5)  # Returns 125
round_nearest(c(2, 7), 5)  # Returns c(0, 5)
}
