% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{round_number}
\alias{round_number}
\title{Round Number}
\usage{
round_number(x, digits = 0)
}
\arguments{
\item{x}{A numeric vector of values to round.}

\item{digits}{Integer indicating the number of decimal places to round to.
The default is 0, indicating rounding to the nearest whole number.}
}
\value{
A numeric vector with the rounded values.
}
\description{
Rounds a numeric vector mathematically, i.e. .5 is rounded up, to a specified
number of decimal places. This is in contrast to the base round function
which follows the IEC 60559 standard and rounds to the even digit in these
cases and sometimes catches users off-guard.
}
\examples{
round_number(1.55, 1)  # Returns 1.6
round_number(c(-1.55, 2.55), 1)  # Returns c(-1.6, 2.6)
}
